# docs/swagger.yml
openapi: "3.0.3"
info:
  title: E-Flight Ticket Team IX
  version: "1.0"
servers:
  - url: https://api.eflight.web.id
  - url: http://localhost:3000
paths:
...
  /api/profile:
    get:
      summary: Mendapatkan profil pengguna berdasarkan ID
      security:
        - BearerAuth: []
        - CookieAuth: []
      operationId: getProfile
      tags:
        - Profile
      responses:
        '200':
          description: Profil pengguna berhasil diambil
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  name:
                    type: string
                    example: "John Doe"
                  email:
                    type: string
                    example: "johndoe@example.com"
                  phoneNumber:
                    type: string
                    example: "081234567890"
                  createdAt:
                    type: string
                    format: date-time
                    example: "2023-12-01T10:00:00Z"
        '404':
          description: Pengguna tidak ditemukan
        '500':
          description: Terjadi kesalahan pada server
    put:
      summary: Memperbarui profil pengguna berdasarkan ID
      security:
        - BearerAuth: []
        - CookieAuth: []
      operationId: updateProfile
      tags:
        - Profile
      requestBody:
        description: Data untuk memperbarui profil pengguna
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  example: "johndoe@example.com"
                phoneNumber:
                  type: string
                  example: "081234567890"
      responses:
        '200':
          description: Profil pengguna berhasil diperbarui
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Profile updated successfully"
                  updatedUser:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: "John Doe"
                      email:
                        type: string
                        example: "johndoe@example.com"
                      phoneNumber:
                        type: string
                        example: "081234567890"
        '404':
          description: Pengguna tidak ditemukan
        '500':
          description: Terjadi kesalahan pada server
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token or connect.sid
      description: Session cookie for authentication